WIP - Add test and refactor test cert creations
This refactors the creation of certs in the tests into a setup method.
This is to dry up the code rather then repeating it all in an extra
test.

The extra test asks the controller to create a new metadata resource but
passes in the sign and enc certs. The test passes at the moment but I'm
not sure why. It checks the values of the generated secret, including
HSM values. The secret shouldn't contain these in this case.

There is some extra code in the original test as well as the refactored
certs method to remove objects that we've created, this is to stop them
clashing with the new test.

It's also worth noting that the original test doesn't specify a
`SAMLSigningCertificate` in the MetadataSpec and it probably should. The
`generateMetadataSecretData` method on the controller should blow up
because of this (as it also doesn't have the certs passed in) but it
doesn't seem to. I'm not sure why.